Broke file into smaller submodules.
