+++ metaserver/renter.go
@@ -1,0 +1,90 @@
+package metaserver
+
+import (
+	"encoding/json"
+	"net/http"
+	"skybin/core"
+	"strconv"
+
+	"github.com/gorilla/mux"
+)
+
+var renters []core.Renter
+
+var postRenterHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	var renter core.Renter
+	_ = json.NewDecoder(r.Body).Decode(&renter)
+	renter.ID = strconv.Itoa(len(renters) + 1)
+	renters = append(renters, renter)
+	json.NewEncoder(w).Encode(renter)
+	dumpDbToFile(providers, renters)
+})
+
+var getRenterHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+	for _, item := range renters {
+		if item.ID == params["id"] {
+			json.NewEncoder(w).Encode(item)
+			return
+		}
+	}
+	w.WriteHeader(http.StatusNotFound)
+})
+
+var getRenterFilesHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+	for _, item := range renters {
+		if item.ID == params["id"] {
+			json.NewEncoder(w).Encode(item.Files)
+			return
+		}
+	}
+	w.WriteHeader(http.StatusNotFound)
+})
+
+var postRenterFileHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+	for i, item := range renters {
+		if item.ID == params["id"] {
+			var file core.File
+			_ = json.NewDecoder(r.Body).Decode(&file)
+			renters[i].Files = append(item.Files, file)
+			json.NewEncoder(w).Encode(item.Files)
+			dumpDbToFile(providers, renters)
+			return
+		}
+	}
+	w.WriteHeader(http.StatusNotFound)
+})
+
+var getRenterFileHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+	for _, item := range renters {
+		if item.ID == params["id"] {
+			for _, file := range item.Files {
+				if file.ID == params["fileId"] {
+					json.NewEncoder(w).Encode(file)
+					return
+				}
+			}
+		}
+	}
+	w.WriteHeader(http.StatusNotFound)
+})
+
+var deleteRenterFileHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+	for _, item := range renters {
+		if item.ID == params["id"] {
+			for i, file := range item.Files {
+				if file.ID == params["fileId"] {
+					item.Files = append(item.Files[:i], item.Files[i+1:]...)
+					json.NewEncoder(w).Encode(file)
+					dumpDbToFile(providers, renters)
+					return
+				}
+			}
+		}
+	}
+	w.WriteHeader(http.StatusNotFound)
+})
